refactor: extract split-hybrid trunk runtime (P3)
Moves the split-hybrid trunk out of the engine monolith into
Sources/RealModelInference/SplitHybridRuntime.swift via a frozen engine
extension (~1900 lines):

- ensureCompiled / ensureHybridCompiled / compileHybridLayers
- generateIncrementalHybrid (+Speculative, +Llama)
- cachedSpeculativeRuntimePair + speculative cache plumbing

Rode-along shared helpers (encodePrompt, composeEmbeddingInput,
writeIncrementalEmbeddingLlama, loadConfigFromMetadataFile,
resolveExactTwoTokenDraft*) land here too; re-homing happens with the
P4 IO pass. ~40 engine members widened private->internal; no behavior
change.

Verified: golden replay bit-identical on all four trunks; full suite green.
